renamed predicates to be more accurate
