+++ controller/predicates/has_request_errors.js
