+++ controller/predicates/has_response_data.js
