+++ test/unit/controller/predicates/has_request_errors.js
@@ -1,13 +1,13 @@
 'use strict';
 
 const _ = require('lodash');
-const has_errors = require('../../../../controller/predicates/has_errors');
+const has_request_errors = require('../../../../controller/predicates/has_request_errors');
 
 module.exports.tests = {};
 
 module.exports.tests.interface = (test, common) => {
   test('valid interface', (t) => {
-    t.equal(typeof has_errors, 'function', 'has_errors is a function');
+    t.equal(typeof has_request_errors, 'function', 'has_request_errors is a function');
     t.end();
   });
 };
@@ -19,7 +19,7 @@ module.exports.tests.true_conditions = (test, common) => {
     };
     const res = {};
 
-    t.ok(has_errors(req, res));
+    t.ok(has_request_errors(req, res));
     t.end();
 
   });
@@ -31,7 +31,7 @@ module.exports.tests.false_conditions = (test, common) => {
     const req = {};
     const res = {};
 
-    t.notOk(has_errors(req, res));
+    t.notOk(has_request_errors(req, res));
     t.end();
 
   });
@@ -42,7 +42,7 @@ module.exports.tests.false_conditions = (test, common) => {
     };
     const res = {};
 
-    t.notOk(has_errors(req, res));
+    t.notOk(has_request_errors(req, res));
     t.end();
 
   });
@@ -51,7 +51,7 @@ module.exports.tests.false_conditions = (test, common) => {
 
 module.exports.all = (tape, common) => {
   function test(name, testFunction) {
-    return tape(`GET /has_errors ${name}`, testFunction);
+    return tape(`GET /has_request_errors ${name}`, testFunction);
   }
 
   for( const testCase in module.exports.tests ){
