+++ test/unit/run.js
@@ -15,8 +15,8 @@ var tests = [
   require('./controller/index'),
   require('./controller/place'),
   require('./controller/search'),
-  require('./controller/predicates/has_data'),
-  require('./controller/predicates/has_errors'),
+  require('./controller/predicates/has_response_data'),
+  require('./controller/predicates/has_request_errors'),
   require('./controller/predicates/is_coarse_reverse'),
   require('./controller/predicates/is_pip_service_enabled'),
   require('./helper/diffPlaces'),
